refactor: extract useObsidian hook
Moves obsidianConfig, obsidianSyncStatus, obsidianLastSynced state,
obsidianVaultHandleRef, obsidianSyncInProgressRef, obsidianPrevTaskStateRef,
obsidianTasksRef, and obsidianInboxRef refs, plus the config persist
effect into a dedicated useObsidian hook.
